requirements.txt
python-dotenv
openai
anthropic
google-genai
tqdm